==> _notebooks/LLMs/pilots/5_data_exploration.Rmd <==
---
title: "reliability_comparison"
author: "Andrew M. Demetriou"
---

```{r}
library("here")           # file logistics
library("tidyverse")      # code logistics
library("ggridges")       # joy plots
library("cowplot")        # merging plots
library("naniar")         # analyzing missingness
library('wordcloud')      # word cloud plotting
library('RColorBrewer')   # custom color pallette
library('tidytext')       # text mining

pilot_dfs <- readRDS(file = here("_data", "_participant_scores", "wave_2", "pilot_2_dfs.RDS"))

values <- c("POWER", "ACHIEVEMENT", "HEDONISM",  
            "STIMULATION", "SELF", "UNIVERSALISM", 
            "BENEVOLENCE", "TRADITION",  
            "CONFORMITY", "SECURITY")
```

```{r}
pilot_dfs <- lapply(pilot_dfs, function(df){
  
  df %>% mutate(c = case_when(
    c == "Completely Confident" ~ 5, 
    c == "Somewhat Confident" ~ 4, 
    c == "Neither Confident nor Unconfident" ~ 3, 
    c == "Somewhat Unconfident" ~ 2, 
    c == "Completely Unconfident" ~ 1
  ))
})
```

```{r}
# self-rated confidence in scores
plots <- lapply(pilot_dfs, function(df){
  df %>% ggplot(aes(x=c)) +
    geom_histogram(bins = 5) +
    theme_minimal() +
    xlab("Confidence")
})

plots[['lyrics']]

ggsave(here("images", "participant_confidence.png"))
```

```{r}
pilot_dfs <- lapply(pilot_dfs, function(df){
  
  df %>% mutate(subjective = case_when(
    `subjective ` == "Completely subjective" ~ 7, 
    `subjective ` == "Very subjective" ~ 6, 
    `subjective ` == "Somewhat subjective" ~ 5, 
    `subjective ` == "Neither subjective nor objective" ~ 4, 
    `subjective ` == "Somewhat objective" ~ 3, 
    `subjective ` == "Very objective" ~ 2, 
    `subjective ` == "Completely objective" ~ 1
  ))
})
```

```{r}
plots <- lapply(pilot_dfs, function(df){
  df %>% ggplot(aes(x=subjective)) +
    geom_histogram(bins = 7) +
    theme_minimal() +
    xlab("Subjectivity")
})

plots[['lyrics']]

ggsave(here("images", "participant_subjectivty.png"))
```






```{r}
df <- pilot_dfs[[1]] %>% 
  select(participant_ID, 
         Q1_1, # I prefer music that contains lyrics over music that does not
         Q1_2, # My favorite songs all have lyrics
         Q1_3, # I skip songs that don't have lyrics
         Q1_4, # If a song has lyrics that I don't like, I skip it
         Q1_5, # I listen to music in languages that I dont' speak
         Q1_6, # I look at the lyrics of songs in the apps of streaming services that I use
         Q2_1  # percentage of music lib
         ) %>%
  unique()
```

```{r}
df <- df %>% 
  select(-Q2_1) %>%
  pivot_longer(cols = paste0('Q1_', seq(1, 6)), 
               names_to = 'question', 
               values_to = 'response') %>%
  mutate(response = case_when(
    response == "Strongly agree" ~ 7, 
    response == "Agree" ~ 6, 
    response == "Somewhat agree" ~ 5, 
    response == "Neither agree nor disagree" ~ 4, 
    response == "Somewhat disagree" ~ 3, 
    response == "Disagree" ~ 2, 
    response == "Strongly disagree" ~ 1, 
  )) %>%
  mutate(question = case_when(
    question == 'Q1_1' ~ "I prefer music that contains lyrics over music that does not", 
    question == 'Q1_2' ~ "My favorite songs all have lyrics", 
    question == 'Q1_3' ~ "I skip songs that don't have lyrics", 
    question == 'Q1_4' ~ "If a song has lyrics that I don't like, I skip it", 
    question == 'Q1_5' ~ "I listen to music in languages that I dont' speak",
    question == 'Q1_6' ~ "I look at the lyrics of songs in the apps of streaming services that I use"
  ))
```

```{r}
df %>%
  ggplot(aes(x=response, y=question, fill = question, alpha = .7)) +
    geom_density_ridges() +
    theme_minimal() +
    scale_x_continuous(breaks = seq(1, 7, 1)) +
    scale_fill_viridis_d(option = "magma") +
    scale_y_discrete(labels = function(x) str_wrap(x, width = 35)) +
    ylab("") + xlab("Response") +
    theme(legend.position="none", axis.text.y = element_text(size = 12))

ggsave(here("images", "participant_survey.png"))
```
